requirements.txt
discord.py
google-generativeai
python-dotenv
yt_dlp
pynacl